+++ sandbox/public/index.html
@@ -12,7 +12,7 @@
     <script src="https://cdn.jsdelivr.net/npm/promise-polyfill@8/dist/polyfill.min.js"></script>
 
     <title>Mock website hosting Alloy</title>
-    <script src="onBeforeEventSentHelper.js" />
+    <script src="/onBeforeEventSentHelper.js"></script>
 
     <!-- prettier-ignore -->
     <script nonce="%REACT_APP_NONCE%">
